refact: add error route and export all routes
