+++ src/server/routes/index.js
@@ -1,0 +1,13 @@
+/**
+ * Copyright (c) 2018-present, NDLA.
+ *
+ * This source code is licensed under the GPLv3 license found in the
+ * LICENSE file in the root directory of this source tree.
+ *
+ */
+
+export { errorRoute } from './errorRoute';
+export { defaultRoute } from './defaultRoute';
+export { iframeArticleRoute } from './iframeArticleRoute';
+export { oembedArticleRoute } from './oembedArticleRoute';
+export { forwardingRoute } from './forwardingRoute';
